+++ client/jetbrains/src/main/java/com/sourcegraph/config/SettingsChangeListener.java
@@ -18,6 +18,8 @@ import com.sourcegraph.cody.statusbar.CodyAutocompleteStatus;
 import com.sourcegraph.cody.statusbar.CodyAutocompleteStatusService;
 import com.sourcegraph.find.browser.JavaToJSBridge;
 import com.sourcegraph.telemetry.GraphQlLogger;
+import com.sourcegraph.utils.CollectionUtil;
+import java.util.List;
 import java.util.Objects;
 import org.jetbrains.annotations.NotNull;
 
@@ -116,6 +118,15 @@ public class SettingsChangeListener implements Disposable {
               ConfigUtil.getAllEditors()
                   .forEach(AutocompleteRenderUtils::rerenderAllAutocompleteInlays);
             }
+
+            // clear autocomplete inlays for blacklisted language editors
+            List<String> languageIdsToClear =
+                CollectionUtil.Companion.diff(
+                    context.newBlacklistedAutocompleteLanguageIds,
+                    context.oldBlacklistedAutocompleteLanguageIds);
+            if (!languageIdsToClear.isEmpty())
+              CodyAutocompleteManager.getInstance()
+                  .clearAutocompleteSuggestionsForLanguageIds(languageIdsToClear);
           }
         });
   }